+++ lib/app/app_widget.dart
@@ -1,6 +1,7 @@
-import 'package:cleanapp/app/controllers/app_controller.dart';
+import 'package:cleanapp/app/app_controller.dart';
 import 'package:cleanapp/app/pages/home/home_page.dart';
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 
 class AppWidget extends StatelessWidget {
 
@@ -10,13 +11,13 @@ class AppWidget extends StatelessWidget {
     return ValueListenableBuilder<bool>(
       valueListenable: AppController.instance.themeSwitch,
       builder: (context, isDark, child)  {
-        return MaterialApp(
+        return GetMaterialApp(
           debugShowCheckedModeBanner: false,
           title: 'Flutter Demo',
           theme: ThemeData(
             primarySwatch: Colors.blue,
             visualDensity: VisualDensity.adaptivePlatformDensity,
-            brightness: isDark?Brightness.dark : Brightness.light,
+            brightness: isDark ? Brightness.dark : Brightness.light,
           ),
           home: HomePage(),
         );
