+++ lib/app/models/appconfig_model.dart
@@ -1,0 +1,6 @@
+import 'package:flutter/material.dart';
+
+class AppConfigModel {
+  final themeSwitch = ValueNotifier<bool>(false);
+
+}
