+++ lib/app/models/user_model.dart
@@ -1,0 +1,8 @@
+class UserModel{
+  final String name;
+  final String email;
+  final String nickname;
+  final String picture;
+
+  UserModel({this.name, this.email, this.nickname, this.picture});
+}
