+++ lib/app/pages/home/components/custom_switch_widget.dart
@@ -1,11 +1,12 @@
-import 'package:cleanapp/app/controllers/app_controller.dart';
+
+import 'package:cleanapp/app/app_controller.dart';
 import 'package:flutter/material.dart';
 
 class CustomSwitchWidget extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Switch(
-        value: AppController.instance.themeSwitch.value,
+        value: AppController.instance.isDark,
         onChanged: (value) {
           AppController.instance.changeTheme(value);
         });
